version 2.0.3 -Rearrange project structure -bug fix -move system changed

+++ src/MazeGame/weapons/BrokenGun.java
@@ -35,7 +35,7 @@ public class BrokenGun extends Weapon {
 
     // ability is fire two fire bullet with greater damage and speed
     @Override
-    protected void cast(int x, int y, int xDest, int yDest) {
+    protected void cast(double x, double y, double xDest, double yDest) {
         double dist = Math.sqrt((xDest - x) * (xDest - x) + (yDest - y) * (yDest - y));
         double xDir = (xDest - x) / dist;
         double yDir = (yDest - y) / dist;
