+++ src/MazeGame/weapons/Rocket.java
@@ -31,7 +31,7 @@ public class Rocket extends Weapon {
     }
 
     @Override
-    protected void cast(int x, int y, int xDest, int yDest) {
+    protected void cast(double x, double y, double xDest, double yDest) {
         double dist = Math.sqrt((xDest - x) * (xDest - x) + (yDest - y) * (yDest - y));
         double xDir = (xDest - x) / dist;
         double yDir = (yDest - y) / dist;
